refactor: REST client abstraction and interfaces
Unify and simplify the REST client layer by replacing ICustomRestClient with a new IRestClient interface, introducing IServiceRestClient for service-specific clients, and renaming CommonRestClientBase to ServiceRestClientBase. Update all service clients to use the new abstractions, improving separation of concerns and code clarity.
