+++ src/Api.Rest/HttpClient/Health/IHealthServiceRestClient.cs
@@ -23,14 +23,14 @@ namespace Zeiss.PiWeb.Api.Rest.HttpClient.Health
 	/// <summary>
 	/// Web service interface to communicate with the REST based PiWeb health service.
 	/// </summary>
-	public interface IHealthServiceRestClient : IRestClient
+	public interface IHealthServiceRestClient : IServiceRestClient
 	{
 		#region properties
 
 		/// <summary>
 		/// A custom rest client that can be used to execute rest request created by a rest request builder.
 		/// </summary>
-		public ICustomRestClient CustomRestClient { get; }
+		public IRestClient RestClient { get; }
 
 		#endregion
